# hand written IstioEndpoint deepcopy