@zirain 2024-08-02
should we use `// +k8s:deepcopy-gen=true` go generate it instead of hand writting?