@wulianglongrd 2024-08-02
> should we use `// +k8s:deepcopy-gen=true` to generate it instead of hand writting?

I am not familiar with this usage, can it be used for any structure? In istio, it is common to write DeepCopy by hand.